Docker-Bot/requirements.txt
discord.py
yt-dlp
requests
PyNaCl
openai